third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-duration-changed-expected.txt

CONSOLE WARNING: 'WebKitMediaSource' is deprecated. Please use 'MediaSource' instead.
Tests duration attribute on MediaSource object

Test setting the duration when closed.
Got expected exception InvalidStateError: Failed to set the 'duration' property on 'WebKitMediaSource': The MediaSource's readyState is not 'open'.
Test getting the duration when closed.

running testExpectExceptionOnNegativeDuration
EVENT(webkitsourceopen)
Got expected exception InvalidAccessError: Failed to set the 'duration' property on 'WebKitMediaSource': The duration provided (-1) is less than the minimum bound (0).

running testExpectExceptionOnNaNDuration
EVENT(webkitsourceopen)
Got expected exception InvalidAccessError: Failed to set the 'duration' property on 'WebKitMediaSource': The duration is not a number.

running testDurationChange
EVENT(webkitsourceopen)
EVENT(loadedmetadata)
EVENT(durationchange)

running testDurationChangeThenOverwriteDuration
EVENT(webkitsourceopen)
EVENT(loadedmetadata)
EVENT(durationchange)
EVENT(durationchange)

running testDurationChangeTruncates
EVENT(webkitsourceopen)
EVENT(loadedmetadata)
EVENT(durationchange)
END OF TEST
